src/server/requirements.txt
Flask
Flask-SQLAlchemy
Gunicorn
mysqlclient
